+++ src/lib/components/table/contracts/ContractsTableRowCTA.tsx
@@ -119,7 +119,7 @@ export const ContractsTableRowCTA = ({
                     />
                   }
                 >
-                  Add or remove from other lists
+                  Add or remove from lists
                 </MenuItem>
               }
             />