+++ src/lib/services/assetService.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { useBaseApiRoute } from "lib/app-provider";
+import { CELATONE_QUERY_KEYS, useBaseApiRoute } from "lib/app-provider";
 import {
   getAssetInfos,
   getAssetInfosWithoutPricesPath,
@@ -20,7 +20,7 @@ export const useAssetInfos = ({
   const assetsApiRoute = useBaseApiRoute("assets");
   const fetchFn = withPrices ? getAssetInfos : getAssetInfosWithoutPricesPath;
   const { data: assets, isLoading } = useQuery(
-    ["query", "assetInfos", assetsApiRoute],
+    [CELATONE_QUERY_KEYS.ASSET_INFOS, assetsApiRoute, withPrices],
     async () => fetchFn(assetsApiRoute),
     { enabled: Boolean(assetsApiRoute), retry: 1, refetchOnWindowFocus: false }
   );
@@ -37,7 +37,7 @@ export const useAssetInfos = ({
 export const useAssetInfoList = () => {
   const assetsApiRoute = useBaseApiRoute("assets");
   return useQuery(
-    ["query", "assetInfosList", assetsApiRoute],
+    [CELATONE_QUERY_KEYS.ASSET_INFO_LIST, assetsApiRoute],
     async () => getAssetInfosWithoutPricesPath(assetsApiRoute),
     { enabled: Boolean(assetsApiRoute), retry: 1, refetchOnWindowFocus: false }
   );
@@ -47,7 +47,7 @@ export const useNativeTokensInfo = () => {
   const nativeTokensApiRoute = useBaseApiRoute("native_tokens");
 
   return useQuery(
-    ["query", "nativeTokensInfo", nativeTokensApiRoute],
+    [CELATONE_QUERY_KEYS.NATIVE_TOKENS_INFO, nativeTokensApiRoute],
     async () => getAssetInfosWithoutPricesPath(nativeTokensApiRoute),
     {
       enabled: Boolean(nativeTokensApiRoute),