+++ src/lib/services/publicProjectService.ts
@@ -4,6 +4,7 @@ import axios from "axios";
 import { useCallback } from "react";
 
 import {
+  CELATONE_QUERY_KEYS,
   useBaseApiRoute,
   usePublicProjectConfig,
   useWasmConfig,
@@ -51,9 +52,13 @@ export const usePublicProjects = (): UseQueryResult<PublicProjectInfo[]> => {
       );
   }, [projectsApiRoute]);
 
-  return useQuery(["public_project", projectsApiRoute], queryFn, {
-    enabled: projectConfig.enabled,
-  });
+  return useQuery(
+    [CELATONE_QUERY_KEYS.PUBLIC_PROJECTS, projectsApiRoute],
+    queryFn,
+    {
+      enabled: projectConfig.enabled,
+    }
+  );
 };
 
 export const usePublicProjectBySlug = (
@@ -73,9 +78,13 @@ export const usePublicProjectBySlug = (
       }));
   }, [projectsApiRoute, slug]);
 
-  return useQuery(["public_project_by_slug", projectsApiRoute, slug], queryFn, {
-    enabled: Boolean(slug) && projectConfig.enabled,
-  });
+  return useQuery(
+    [CELATONE_QUERY_KEYS.PUBLIC_PROJECT_BY_SLUG, projectsApiRoute, slug],
+    queryFn,
+    {
+      enabled: Boolean(slug) && projectConfig.enabled,
+    }
+  );
 };
 
 export const usePublicProjectByContractAddress = (
@@ -96,7 +105,11 @@ export const usePublicProjectByContractAddress = (
   }, [projectsApiRoute, contractAddress]);
 
   return useQuery(
-    ["public_project_by_contract_address", projectsApiRoute, contractAddress],
+    [
+      CELATONE_QUERY_KEYS.PUBLIC_PROJECT_BY_CONTRACT_ADDRESS,
+      projectsApiRoute,
+      contractAddress,
+    ],
     queryFn,
     {
       enabled:
@@ -124,7 +137,7 @@ export const usePublicProjectByCodeId = (
   }, [projectsApiRoute, codeId]);
 
   return useQuery(
-    ["public_project_by_code_id", projectsApiRoute, codeId],
+    [CELATONE_QUERY_KEYS.PUBLIC_PROJECT_BY_CODE_ID, projectsApiRoute, codeId],
     queryFn,
     {
       enabled: isCodeId(codeId) && projectConfig.enabled && wasmConfig.enabled,
@@ -150,7 +163,11 @@ export const usePublicProjectByAccountAddress = (
       .then(({ data: projectInfo }) => projectInfo);
   }, [accountAddress, projectsApiRoute]);
   return useQuery(
-    ["public_project_by_account_address", projectsApiRoute, accountAddress],
+    [
+      CELATONE_QUERY_KEYS.PUBLIC_PROJECT_BY_WALLET_ADDRESS,
+      projectsApiRoute,
+      accountAddress,
+    ],
     queryFn,
     {
       enabled: Boolean(accountAddress) && projectConfig.enabled,